ref(ref-react-16): Refactor the homepage table to use React 16 fragments
Refactor the homepage to use react 16 fragments and componitize it more
