+++ src/routes.js
@@ -1,5 +1,5 @@
 //eslint: ignor
-import React from 'react';
+import React, {Fragment} from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
 import {ConnectedRouter} from 'react-router-redux';
 import {Home, Portfolio, Navigation} from './features';
@@ -8,11 +8,11 @@ import {history} from './store';
 export default () => (
 	<ConnectedRouter history={history}>
 		<Router>
-			<div>
+			<Fragment>
 				<Navigation />
 				<Route exact path="/" component={Home} />
 				<Route path="/portfolio" component={Portfolio} />
-			</div>
+			</Fragment>
 		</Router>
 	</ConnectedRouter>
 );
